+++ apps/hyperlocalise-web/src/api/routes/project/job.route.ts
@@ -81,7 +81,7 @@ import {
 import { resolveProviderSourceFiles } from "@/lib/providers/job-provider-source-files";
 import { mapProviderQaErrorToHttpStatus } from "@/lib/providers/map-provider-qa-http-error";
 import { runProviderJobQaForJob } from "@/lib/providers/agent-runs/provider-agent-qa";
-import { getProviderContentPuller } from "@/lib/providers/provider-content-pullers";
+import { getProviderContentPuller } from "@/lib/providers/adapters/tms-provider-adapter-registry";
 import { maybeEnqueueAutoWriteBackAfterProposalReview } from "@/lib/providers/agent-runs/tms-agent-automation-runner";
 import { getActiveOrganizationExternalTmsProviderCredentialRow } from "@/lib/providers/organization-external-tms-provider-credentials";
 import {
