+++ apps/hyperlocalise-web/src/lib/providers/match-resolution.ts
@@ -1,8 +1,10 @@
 import { memorySupportsLiveSearch } from "@/lib/providers/contracts/memory-live-search";
 import type { GlossaryMatchResolution } from "@/lib/providers/contracts/glossary-matcher";
 import type { TranslationMemoryMatchResolution } from "@/lib/providers/contracts/translation-memory-matcher";
-import { getProviderGlossaryMatcher } from "@/lib/providers/provider-glossary-matchers";
-import { getProviderTranslationMemoryMatcher } from "@/lib/providers/provider-translation-memory-matchers";
+import {
+  getProviderGlossaryMatcher,
+  getProviderTranslationMemoryMatcher,
+} from "@/lib/providers/adapters/tms-provider-adapter-registry";
 
 export const defaultTranslationMemoryMatchResolution: TranslationMemoryMatchResolution = {
   getProviderTranslationMemoryMatcher,
