--- apps/hyperlocalise-web/src/lib/providers/provider-comment-pushers.ts
@@ -1,1 +1,0 @@
-export { getProviderCommentPusher } from "@/lib/providers/adapters/tms-provider-adapter-registry";
