--- apps/hyperlocalise-web/src/lib/providers/provider-content-pullers.ts
@@ -1,1 +1,0 @@
-export { getProviderContentPuller } from "@/lib/providers/adapters/tms-provider-adapter-registry";
