--- apps/hyperlocalise-web/src/lib/providers/provider-glossary-matchers.ts
@@ -1,8 +1,0 @@
-import type {
-  ExternalTmsGlossaryMatcher,
-  ExternalTmsGlossaryMatcherInput,
-} from "@/lib/providers/contracts/glossary-matcher";
-
-export type { ExternalTmsGlossaryMatcher, ExternalTmsGlossaryMatcherInput };
-
-export { getProviderGlossaryMatcher } from "@/lib/providers/adapters/tms-provider-adapter-registry";
