--- apps/hyperlocalise-web/src/lib/providers/provider-translation-pushers.ts
@@ -1,1 +1,0 @@
-export { getProviderTranslationPusher } from "@/lib/providers/adapters/tms-provider-adapter-registry";
