--- apps/hyperlocalise-web/src/lib/providers/tms-provider-fetcher-registry.ts
@@ -1,7 +1,0 @@
-export {
-  tmsProviderFileKeyFetchers,
-  tmsProviderGlossaryFetchers,
-  tmsProviderJobTaskFetchers,
-  tmsProviderProjectFetchers,
-  tmsProviderTranslationMemoryFetchers,
-} from "@/lib/providers/adapters/tms-provider-adapter-registry";
